# handshake i/o batching