what the title says 🤷🏽 also minor cleanup

**Before**

```
handshake_test.go
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/btprotocol
cpu: Apple M1 Pro
BenchmarkHandshakeRoundTrip-10        284175          4229 ns/op        1456 B/op         20 allocs/op
PASS
ok      github.com/james-lawrence/torrent/btprotocol    2.315s
```

**After**

```
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/btprotocol
cpu: Apple M1 Pro
BenchmarkHandshakeRoundTrip-10        382554          3125 ns/op        1296 B/op         14 allocs/op
PASS
ok      github.com/james-lawrence/torrent/btprotocol    2.235s
```